+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Delay downloads (e.g.: --delay 3s)
 - Download sourcemaps to `{output}/sourcemaps`
 - Download log added
+- Local sourcemap cache added
 
 ### Breaking changes
 - NaN